Fix cigar reversing
